Module-7: Grouped use cases implementing classes by files.
