+++ core/src/main/kotlin/me/vgolovnin/ddd/delivery/core/application/usecases/AssignOrder.kt
@@ -9,6 +9,8 @@ import me.vgolovnin.ddd.delivery.core.ports.CourierRepository
 import me.vgolovnin.ddd.delivery.core.ports.OrderRepository
 import me.vgolovnin.ddd.delivery.core.utils.UnitOfWork
 
+object AssignOrderCommand
+
 class AssignOrderHandler(
     private val courierRepository: CourierRepository,
     private val orderRepository: OrderRepository,
