+++ core/src/main/kotlin/me/vgolovnin/ddd/delivery/core/application/usecases/CreateOrder.kt
@@ -3,6 +3,12 @@ package me.vgolovnin.ddd.delivery.core.application.usecases
 import me.vgolovnin.ddd.delivery.core.domain.model.order.Order
 import me.vgolovnin.ddd.delivery.core.domain.sharedkernel.Location
 import me.vgolovnin.ddd.delivery.core.ports.OrderRepository
+import java.util.*
+
+data class CreateOrderCommand(
+    val basketId: UUID,
+    val street: String,
+)
 
 class CreateOrderHandler(
     private val orderRepository: OrderRepository,
