--- core/src/main/kotlin/me/vgolovnin/ddd/delivery/core/application/usecases/CreateOrderCommand.kt
@@ -1,8 +1,0 @@
-package me.vgolovnin.ddd.delivery.core.application.usecases
-
-import java.util.*
-
-data class CreateOrderCommand(
-    val basketId: UUID,
-    val street: String,
-)
