--- core/src/main/kotlin/me/vgolovnin/ddd/delivery/core/application/usecases/UpdateCourierLocationsCommand.kt
@@ -1,3 +1,0 @@
-package me.vgolovnin.ddd.delivery.core.application.usecases
-
-object UpdateCourierLocationsCommand
